+++ bcs-app/backend/apps/cluster/views/__init__.py
@@ -43,5 +43,5 @@ from .node_views.query_apis import (
 )
 
 from .node_views.operation_apis import (
-    DeleteNotReadyNode, BatchReinstallNodes, SetNodelabelsViewSets
+    DeleteNotReadyNode, BatchReinstallNodes, NodelabelsViewSets
 )
