+++ bcs-app/backend/apps/cluster/views/node_views/operation_apis.py
@@ -15,13 +15,16 @@ from rest_framework import response, viewsets
 from rest_framework.renderers import BrowsableAPIRenderer
 from django.utils.translation import ugettext_lazy as _
 
+from backend.accounts import bcs_perm
 from backend.activity_log import client
 from backend.utils.errcodes import ErrorCode
 from backend.apps.cluster.views_bk import node
 from backend.utils.renderers import BKAPIRenderer
 from backend.utils.error_codes import error_codes
+from backend.apps.cluster.models import NodeStatus
 from backend.resources.cluster import utils as node_utils
 from backend.resources.project.constants import ProjectKind
+from backend.apps.cluster.utils import cluster_env_transfer
 from backend.apps.cluster.models import CommonStatus, NodeUpdateLog
 from backend.apps.cluster import serializers as node_serializers
 from backend.apps.cluster.views.node_views import serializers as node_slz
@@ -95,7 +98,7 @@ class BatchReinstallNodes(ClusterPerm, Nodes, viewsets.ViewSet):
         return response.Response()
 
 
-class SetNodelabelsViewSets(viewsets.ViewSet):
+class NodelabelsViewSets(viewsets.ViewSet):
     renderer_classes = (BKAPIRenderer, BrowsableAPIRenderer)
 
     def set_mesos_node_labels(self, access_token, project_id, node_labels):
@@ -119,7 +122,7 @@ class SetNodelabelsViewSets(viewsets.ViewSet):
         pass
 
     def set_labels(self, request, project_id):
-        slz = node_slz.CreateNodeLabelsSLZ(data=request.data)
+        slz = node_slz.SetNodeLabelsSLZ(data=request.data)
         slz.is_valid(raise_exception=True)
         data = slz.validated_data
 
@@ -128,3 +131,49 @@ class SetNodelabelsViewSets(viewsets.ViewSet):
             request.user.token.access_token, project_id, data["node_labels"])
 
         return response.Response()
+
+    def _mesos_node_labels(self, access_token, project_id, nodes):
+        # 组装查询数据
+        # 格式 {"cluster-id1": ["ip1", "ip2"], "cluster-id2": ["ip3", "ip4"]]
+        # 节点ip为空时，查询所有节点的labels，避免每个IP匹配一次
+        cluster_nodes = {node["cluster_id"]: [] for node in nodes}
+        return node_utils.query_mesos_node_labels(access_token, project_id, cluster_nodes)
+
+    def _cluster_id_map(self, access_token, project_id):
+        clusters = node_utils.get_clusters(access_token, project_id)
+        return {
+            cluster["cluster_id"]: {
+                "cluster_env": cluster_env_transfer(cluster["environment"]),
+                "cluster_name": cluster["name"]
+            }
+            for cluster in clusters
+        }
+
+    def compose_node_data(self, nodes, labels, cluster_id_map):
+        """组装node数据，添加label、集群名称
+        """
+        for node_info in nodes:
+            inner_ip = node_info["inner_ip"]
+            cluster_id = node_info["cluster_id"]
+            node_info["labels"] = labels.get(cluster_id, {}).get(inner_ip, [])
+            node_info.update(cluster_id_map.get(cluster_id, {}))
+
+    def list(self, request, project_id):
+        # TODO: 现阶段仅针对mesos
+        if request.project.kind == ProjectKind.K8S.value:
+            raise error_codes.NotOpen()
+
+        access_token = request.user.token.access_token
+        cluster_id = request.query_params.get("cluster_id")
+        # cluster_id 为None时，查询项目下的所有集群的节点
+        nodes = node_utils.get_cluster_nodes(access_token, project_id, cluster_id, raise_exception=False)
+        # 排除状态为已删除的节点
+        nodes = [node for node in nodes if node["status"] not in [NodeStatus.Removed]]
+        project_kind_name = ProjectKind.get_choice_label(request.project.kind)
+        labels = getattr(self, f"_{project_kind_name.lower()}_node_labels")(access_token, project_id, nodes)
+        cluster_id_map = self._cluster_id_map(access_token, project_id)
+        self.compose_node_data(nodes, labels, cluster_id_map)
+        # 添加权限
+        nodes_results = bcs_perm.Cluster.hook_perms(request, project_id, nodes)
+
+        return response.Response({'count': len(nodes_results), 'results': nodes_results})
