+++ bcs-app/backend/apps/cluster/views/node_views/query_apis.py
@@ -18,13 +18,11 @@ from rest_framework import response, viewsets
 from rest_framework.renderers import BrowsableAPIRenderer
 from django.http import HttpResponse
 
-from backend.accounts import bcs_perm
 from backend.components import paas_cc
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
 from backend.utils.renderers import BKAPIRenderer
 from backend.resources.cluster import utils as node_utils
-from backend.apps.cluster.utils import cluster_env_transfer
 from backend.resources.project.constants import ProjectKind
 from backend.apps.cluster import constants as node_constants
 from backend.apps.cluster.models import NodeLabel, NodeStatus
@@ -145,7 +143,7 @@ class ListNodelabelsViewSets(viewsets.ViewSet):
         node_labels = node_utils.query_mesos_node_labels(access_token, project_id, data)
         return node_labels.values()
 
-    def query_k8s_node_labels(self):
+    def query_k8s_node_labels(self, access_token, project_id, data):
         pass
 
     def list_labels(self, request, project_id):
@@ -166,49 +164,3 @@ class ListNodelabelsViewSets(viewsets.ViewSet):
             request.user.token.access_token, project_id, cluster_nodes)
 
         return response.Response(labels)
-
-    def _mesos_node_labels(self, access_token, project_id, nodes):
-        # 组装查询数据
-        # 格式 {"cluster-id1": ["ip1", "ip2"], "cluster-id2": ["ip3", "ip4"]]
-        # 节点ip为空时，查询所有节点的labels，避免每个IP匹配一次
-        cluster_nodes = {node["cluster_id"]: [] for node in nodes}
-        return node_utils.query_mesos_node_labels(access_token, project_id, cluster_nodes)
-
-    def _cluster_id_map(self, access_token, project_id):
-        clusters = node_utils.get_clusters(access_token, project_id)
-        return {
-            cluster["cluster_id"]: {
-                "cluster_env": cluster_env_transfer(cluster["environment"]),
-                "cluster_name": cluster["name"]
-            }
-            for cluster in clusters
-        }
-
-    def compose_node_data(self, nodes, labels, cluster_id_map):
-        """组装node数据，添加label、集群名称
-        """
-        for node in nodes:
-            inner_ip = node["inner_ip"]
-            cluster_id = node["cluster_id"]
-            node["labels"] = labels.get(cluster_id, {}).get(inner_ip, [])
-            node.update(cluster_id_map.get(cluster_id, {}))
-
-    def list(self, request, project_id):
-        # TODO: 现阶段仅针对mesos
-        if request.project.kind == ProjectKind.K8S.value:
-            raise error_codes.NotOpen()
-
-        access_token = request.user.token.access_token
-        cluster_id = request.query_params.get("cluster_id")
-        # cluster_id 为None时，查询项目下的所有集群的节点
-        nodes = node_utils.get_cluster_nodes(access_token, project_id, cluster_id, raise_exception=False)
-        # 排除状态为已删除的节点
-        nodes = [node for node in nodes if node["status"] not in [NodeStatus.Removed]]
-        project_kind_name = ProjectKind.get_choice_label(request.project.kind)
-        labels = getattr(self, f"_{project_kind_name.lower()}_node_labels")(access_token, project_id, nodes)
-        cluster_id_map = self._cluster_id_map(access_token, project_id)
-        self.compose_node_data(nodes, labels, cluster_id_map)
-        # 添加权限
-        nodes_results = bcs_perm.Cluster.hook_perms(request, project_id, nodes)
-
-        return response.Response({'count': len(nodes_results), 'results': nodes_results})
